# docs/migration/openapi-step5.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AlphaQuant API",
    "description": "迁移阶段本机接口；支持持久化后台任务和同步兼容接口。",
    "version": "0.3.0"
  },
  "paths": {
...
    "/api/v1/research/plans": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Plans",
        "operationId": "plans_api_v1_research_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Save Plan",
        "operationId": "save_plan_api_v1_research_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/plans/{plan_id}/{revision}/request": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Plan Request",
        "operationId": "plan_request_api_v1_research_plans__plan_id___revision__request_get",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "revision",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Revision"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/plans/compare": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Compare Plans",
        "operationId": "compare_plans_api_v1_research_plans_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanCompareInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audits/external/preview": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "External Preview",
        "operationId": "external_preview_api_v1_audits_external_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audits/external/check": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "External Check",
        "operationId": "external_check_api_v1_audits_external_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/backfill/jobs": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Backfill Records",
        "operationId": "backfill_records_api_v1_data_backfill_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Start Backfill",
        "operationId": "start_backfill_api_v1_data_backfill_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackfillInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/backfill/jobs/{identifier}/stop": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Stop Backfill",
        "operationId": "stop_backfill_api_v1_data_backfill_jobs__identifier__stop_post",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/backfill/jobs/{identifier}/retry": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Retry Backfill",
        "operationId": "retry_backfill_api_v1_data_backfill_jobs__identifier__retry_post",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/data/backfill/jobs/{identifier}/log": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Backfill Log",
        "operationId": "backfill_log_api_v1_data_backfill_jobs__identifier__log_get",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackfillInput": {
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "datasets": {
            "items": {
              "type": "string",
              "enum": [
                "bars",
                "actions",
                "derived"
              ]
            },
            "type": "array",
            "minItems": 1,
            "title": "Datasets"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "datasets"
        ],
        "title": "BackfillInput"
      },
      "BacktestInput": {
        "properties": {
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "snapshot_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Run Id"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "plan_revision": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Revision"
          },
          "strategy_plugin": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Strategy Plugin"
          },
          "strategy_id": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Strategy Id"
          },
          "strategy_parameters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "string"
                }
              ]
            },
            "type": "object",
            "title": "Strategy Parameters"
          },
          "initial_cash": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Initial Cash"
          },
          "top_n": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Top N"
          },
          "rebalance": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "monthly"
            ],
            "title": "Rebalance"
          },
          "portfolio_method": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "equal_weight",
                  "inverse_volatility",
                  "risk_parity",
                  "mean_variance"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Method"
          },
          "universe_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "fixed",
                  "historical"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Universe Mode"
          },
          "risk_limits": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskInput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "start_date",
          "end_date",
          "strategy_plugin",
          "strategy_id",
          "initial_cash",
          "top_n",
          "rebalance"
        ],
        "title": "BacktestInput"
      },
...
      "ExternalInput": {
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 5000000,
            "minLength": 1,
            "title": "Content"
          },
          "mapping": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapping"
          },
          "unit": {
            "type": "string",
            "enum": [
              "股",
              "手"
            ],
            "title": "Unit",
            "default": "股"
          },
          "raw_prices": {
            "type": "boolean",
            "title": "Raw Prices",
            "default": true
          },
          "duplicate_confirmed": {
            "type": "boolean",
            "title": "Duplicate Confirmed",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ExternalInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanCompareInput": {
        "properties": {
          "versions": {
            "items": {
              "$ref": "#/components/schemas/PlanRef"
            },
            "type": "array",
            "maxItems": 4,
            "minItems": 2,
            "title": "Versions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "versions"
        ],
        "title": "PlanCompareInput"
      },
      "PlanInput": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BacktestInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "inputs": {
            "type": "object",
            "title": "Inputs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "PlanInput"
      },
      "PlanRef": {
        "properties": {
          "plan_id": {
            "type": "string",
            "title": "Plan Id"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan_id",
          "revision"
        ],
        "title": "PlanRef"
      },
...
      "RiskInput": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "max_total_weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Max Total Weight",
            "default": 1.0
          },
          "max_single_weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Max Single Weight",
            "default": 1.0
          },
          "max_positions": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Max Positions",
            "default": 10
          },
          "minimum_cash_ratio": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Minimum Cash Ratio",
            "default": 0
          },
          "max_drawdown": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Max Drawdown",
            "default": 0.2
          },
          "daily_position_limits": {
            "type": "boolean",
            "title": "Daily Position Limits",
            "default": true
          },
          "drawdown_action": {
            "type": "string",
            "enum": [
              "stop_new",
              "reduce",
              "liquidate"
            ],
            "title": "Drawdown Action",
            "default": "stop_new"
          },
          "drawdown_target_weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Drawdown Target Weight",
            "default": 0.5
          },
          "max_industry_weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Max Industry Weight",
            "default": 1
          },
          "max_daily_loss": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Max Daily Loss",
            "default": 1
          },
          "max_rebalance_turnover": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Max Rebalance Turnover",
            "default": 2
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RiskInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}